test-resolve util moved to vitest folder
