+++ packages/crafty-preset-vitest/src/module-directories-plugin.js
@@ -1,8 +1,6 @@
 const path = require("node:path");
 const enhancedResolve = require("@swissquote/crafty/packages/enhanced-resolve.js");
-const {
-  getCraftyTestResolveOptions
-} = require("@swissquote/crafty/packages/test-resolve.js");
+const { getCraftyTestResolveOptions } = require("./test-resolve.js");
 
 function isBareImport(specifier) {
   return (
